JDK-8265012
