+++ src/hotspot/share/gc/shenandoah/shenandoahMark.inline.hpp
@@ -40,6 +40,11 @@
 #include "utilities/powerOfTwo.hpp"
 
 template <class T>
+void ShenandoahInitMarkRootsClosure::do_oop_work(T* p) {
+  ShenandoahMark::mark_through_ref<T, NO_DEDUP>(p, _queue, _mark_context, false);
+}
+
+template <class T>
 void ShenandoahMark::do_task(ShenandoahObjToScanQueue* q, T* cl, ShenandoahLiveData* live_data, ShenandoahMarkTask* task) {
   oop obj = task->obj();
